clustering_requirements.txt
numpy==1.26.4
pandas==2.2.2
Pillow==10.4.0
hdbscan==0.8.37
torch==2.3.1
umap-learn==0.5.6
torchvision==0.18.1